server/requirements.txt
deta==1.1.0
fastapi==0.92.0
jose==1.0.0
passlib==1.7.4
pydantic==1.10.4
python-dotenv==1.0.0
python_bcrypt==0.3.2
python_jose==3.3.0
uvicorn==0.20.0
bcrypt
pydantic[email]
python-multipart
